add: windNoteの削除と編集をuserIdが一致する時だけ実行できるように

+++ app/UseCases/WindNote/WindNoteDeleteAction.php
@@ -3,11 +3,20 @@
 namespace App\UseCases\WindNote;
 
 use App\Models\WindNote;
+use Illuminate\Support\Facades\Auth;
 
 class WindNoteDeleteAction
 {
-    public function __invoke($windNote)
+    public function __invoke(WindNote $windNote)
     {
+        $user = Auth::user();
+
+        if ($user->id !== $windNote->user_id) {
+            return response()->json([
+                'message' => '削除する権限がありません',
+            ], 403);
+        }
+
         $windNote->delete();
 
         return response()->json([
